requirements-dev.txt
# Tests
numpy
pytest
pytest-cov
pytest-xdist
pytest-timeout
dill
filelock
setuptools
# Test contrib dependencies
scipy
pytorch_fid
tqdm
scikit-learn
matplotlib
tensorboardX
visdom
polyaxon
wandb
mlflow
neptune-client>=0.16.17
tensorboard
torchvision
pynvml
clearml
scikit-image
py-rouge
pycocotools
# temporary fix for python=3.12 and v3.8.1
# nltk
git+https://github.com/nltk/nltk
# Examples dependencies
pandas
gymnasium
# temporary fix: E   AttributeError: module 'mpmath' has no attribute 'rational'
mpmath<1.4
